requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.25.1
langchain==0.0.335
langchain-openai==0.0.2
openai==1.3.3
redis==5.0.1
pyjwt==2.8.0
pydantic==2.4.2